requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.2.1
docopt==0.6.2
et-xmlfile==1.1.0
eventlet==0.37.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.14.0
Flask==3.0.3
Flask-SocketIO==5.4.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
gTTS==2.5.3
gunicorn==23.0.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
intel-openmp==2021.4.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jira==3.8.0
joblib==1.3.2
Js2Py==0.74
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==3.3.3
kiwisolver==1.4.5
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.38
langchain-nvidia-ai-endpoints==0.2.2
langchain-text-splitters==0.2.1
langsmith==0.1.75
libclang==16.0.6
llvmlite==0.43.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mixupy==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.2.1
neurokit2==0.2.7
notebook==7.2.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
oauthlib==3.2.2
openai==1.32.0
openai-whisper @ git+https://github.com/openai/whisper@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
pipwin==0.5.2
platformdirs==3.11.0
playsound==1.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.16.1
pyjsparser==2.7.1
pyparsing==3.1.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.8.2
python-engineio==4.10.0
python-json-logger==2.0.7
python-multipart==0.0.10
python-socketio==5.11.4
pytz==2023.4
pywebio==1.8.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.6.7
safetensors==0.4.3
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.38.6
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.16.0
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.3.1
tornado==6.3.3
tqdm==4.66.4
traitlets==5.13.0
transformers==4.41.2
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
tzlocal==5.2
ua-parser==0.18.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
user-agents==2.2.0
uvicorn==0.30.6
vboxapi==1.0
wcwidth==0.2.9
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.11
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.17.0
